Move storage quota to a separate package as preparation for quota interfaces.
