+++ ambry-api/src/main/java/com/github/ambry/quota/storage/QuotaOperation.java
@@ -11,11 +11,11 @@
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  */
-package com.github.ambry.quota;
+package com.github.ambry.quota.storage;
 
 /**
- * Quota service mode. If mode is tracking, it will not throttle traffics even if the quota is exceeded.
+ * Operations for quota service.
  */
-public enum QuotaMode {
-  Tracking, Throttling
+public enum QuotaOperation {
+  Post, Delete, TtlUpdate
 }
