+++ ambry-frontend/src/main/java/com/github/ambry/frontend/FrontendRestRequestService.java
@@ -22,7 +22,7 @@ import com.github.ambry.commons.Callback;
 import com.github.ambry.config.FrontendConfig;
 import com.github.ambry.messageformat.BlobInfo;
 import com.github.ambry.protocol.GetOption;
-import com.github.ambry.quota.StorageQuotaService;
+import com.github.ambry.quota.storage.StorageQuotaService;
 import com.github.ambry.rest.RequestPath;
 import com.github.ambry.rest.ResponseStatus;
 import com.github.ambry.rest.RestMethod;
