+++ ambry-quota/src/main/java/com/github/ambry/quota/storage/AmbryStorageQuotaService.java
@@ -11,7 +11,7 @@
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  */
-package com.github.ambry.quota;
+package com.github.ambry.quota.storage;
 
 import com.github.ambry.commons.CommonUtils;
 import com.github.ambry.config.StorageQuotaConfig;
