+++ ambry-quota/src/main/java/com/github/ambry/quota/storage/AmbryStorageQuotaServiceFactory.java
@@ -11,16 +11,11 @@
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  */
-package com.github.ambry.quota;
+package com.github.ambry.quota.storage;
 
 import com.codahale.metrics.MetricRegistry;
-import com.github.ambry.commons.CommonUtils;
 import com.github.ambry.config.StorageQuotaConfig;
 import com.github.ambry.config.VerifiableProperties;
-import com.github.ambry.utils.Utils;
-import java.util.concurrent.ScheduledExecutorService;
-import org.apache.helix.store.HelixPropertyStore;
-import org.apache.helix.zookeeper.datamodel.ZNRecord;
 
 
 /**
