+++ ambry-quota/src/test/java/com/github/ambry/quota/storage/AmbryStorageQuotaEnforcerTest.java
@@ -11,7 +11,7 @@
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  */
-package com.github.ambry.quota;
+package com.github.ambry.quota.storage;
 
 import com.github.ambry.utils.TestUtils;
 import java.util.Collections;
